+++ packages/sdk/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the Client code for "@zowe/zowex-for-zowe-sdk" are docume
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## Recent Changes
+
+- Added support to list link list data sets. [#1061](https://github.com/zowe/zowex/pull/1061)
+
 ## `0.6.1`
 
 - Added new functions to the `ZSshUtils` class called `detectServerOnPath()` and `lacksWriteAccess()`. [#1055](https://github.com/zowe/zowex/pull/1055)